changed some int types to NSInteger type
git-svn-id: svn://orca.physics.unc.edu/Orca@11744 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Document/ORDocumentController.h
@@ -88,7 +88,7 @@
 - (IBAction) setProductionMode:(id)sender;
 
 #pragma mark ¥¥¥Data Source
-- (id)   outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item;
+- (id)   outlineView:(NSOutlineView *)outlineView child:(NSUInteger)index ofItem:(id)item;
 - (BOOL) outlineView:(NSOutlineView *)outlineView isItemExpandable:(id)item;
 - (int)  outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item;
 - (id)  outlineView:(NSOutlineView *)outlineView  objectValueForTableColumn:(NSTableColumn *)tableColumn byItem:(id)item;
